feat: extract parseInline and apply to PollSlide question
